Move scale transforms into nested module
